# backend/openapi.yaml
openapi: 3.1.0
...
info:
  title: LoopBuy Backend API
  version: 1.0.0
  description: >-
    REST API for LoopBuy accounts, marketplace listings, social features,
    scam screening, recommendations, and the RAG shopping assistant.
servers:
  - url: http://localhost:8090
    description: Docker Compose local API
security: []
...
paths:
...
  /api/v1/users/me/favourites:
    get:
      tags: [Favourites]
      operationId: listFavourites
      summary: List saved listings
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/FavouritesLimit'
      responses:
        '200':
          description: Saved active and approved listings.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListingList'
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/users/me/favourites/{listingId}:
    parameters:
      - $ref: '#/components/parameters/ListingId'
    put:
      tags: [Favourites]
      operationId: addFavourite
      summary: Save a listing idempotently
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Listing is saved.
        default:
          $ref: '#/components/responses/Problem'
    delete:
      tags: [Favourites]
      operationId: removeFavourite
      summary: Remove a saved listing
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Listing is no longer saved.
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/users/me/cart:
    get:
      tags: [Cart]
      operationId: getCart
      summary: Get the current user's cart
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Cart with active and approved listing details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/users/me/cart/items/{listingId}:
    parameters:
      - $ref: '#/components/parameters/ListingId'
    put:
      tags: [Cart]
      operationId: putCartItem
      summary: Set a cart item quantity
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartQuantityRequest'
      responses:
        '200':
          description: Updated cart.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        default:
          $ref: '#/components/responses/Problem'
    patch:
      tags: [Cart]
      operationId: patchCartItem
      summary: Set a cart item quantity
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartQuantityRequest'
      responses:
        '200':
          description: Updated cart.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      tags: [Cart]
      operationId: removeCartItem
      summary: Remove a listing from the cart
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Updated cart.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/users/me/cart/items:
    delete:
      tags: [Cart]
      operationId: clearCart
      summary: Clear the current user's cart
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Empty cart.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/conversations:
    post:
      tags: [Conversations]
      operationId: createConversation
      summary: Create or reuse a conversation for a listing
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConversationCreateRequest'
      responses:
        '201':
          description: New or existing buyer/seller conversation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        default:
          $ref: '#/components/responses/Problem'
    get:
      tags: [Conversations]
      operationId: listConversations
      summary: List the current user's conversations
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Conversations ordered by recent activity.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationList'
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/conversations/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Conversations]
      operationId: getConversation
      summary: Get a conversation as a member
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Conversation details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      tags: [Conversations]
      operationId: leaveConversation
      summary: Leave a conversation
      description: Marks the current membership as left; the conversation is retained.
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Membership marked as left.
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/conversations/{id}/messages:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Conversations]
      operationId: listMessages
      summary: List messages in ascending order
      security:
        - bearerAuth: []
      parameters:
        - name: before
          in: query
          description: Return message IDs below this cursor.
          schema:
            type: integer
            format: int64
            minimum: 1
        - $ref: '#/components/parameters/MessagesLimit'
      responses:
        '200':
          description: Conversation messages.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageList'
        default:
          $ref: '#/components/responses/Problem'
    post:
      tags: [Conversations]
      operationId: createMessage
      summary: Send a message
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageRequest'
      responses:
        '201':
          description: Message created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/conversations/{id}/messages/{messageId}:
    parameters:
      - $ref: '#/components/parameters/Id'
      - $ref: '#/components/parameters/MessageId'
    patch:
      tags: [Conversations]
      operationId: updateMessage
      summary: Edit a message sent by the current user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageRequest'
      responses:
        '200':
          description: Updated message.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      tags: [Conversations]
      operationId: deleteMessage
      summary: Delete a message sent by the current user
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Message deleted.
        default:
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: HS256 access token issued by register, login, or refresh.
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
...
    ListingId:
      name: listingId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
...
    MessageId:
      name: messageId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
...
    FavouritesLimit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    MessagesLimit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
...
  responses:
    Problem:
      description: Error response. Typical statuses are 400, 401, 403, 404, 409, 422, 428, 500, or 503.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
        errors:
          type: object
          additionalProperties:
            type: string
...
    CartQuantityRequest:
      type: object
      additionalProperties: false
      required: [quantity]
      properties:
        quantity:
          type: integer
          minimum: 1
          maximum: 99
    ConversationCreateRequest:
      type: object
      additionalProperties: false
      required: [listing_id]
      properties:
        listing_id:
          type: integer
          format: int64
          minimum: 1
    MessageRequest:
      type: object
      additionalProperties: false
      required: [message_text]
      properties:
        message_text:
          type: string
          minLength: 1
          maxLength: 5000
...
    ItemCondition:
      type: string
      enum: [new, like_new, good, fair]
    Profile:
      type: object
      additionalProperties: false
      required: [user_id, full_name, phone, location, bio, profile_image, updated_at]
      properties:
        user_id:
          type: integer
          format: int64
        full_name:
          type: string
        phone:
          type: string
          description: Empty in public-user responses.
        location:
          type: string
        bio:
          type: string
        profile_image:
          type: string
        updated_at:
          type: string
          format: date-time
    User:
      type: object
      additionalProperties: false
      required: [user_id, username, email_verified, created_at]
      properties:
        user_id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
          format: email
          description: Present only in private-user responses.
        email_verified:
          type: boolean
          description: False in public-user responses; authoritative in private/session responses.
        role:
          type: string
          enum: [user, moderator, admin]
          description: Omitted in public-user responses.
        status:
          type: string
          enum: [active, suspended, deleted]
          description: Omitted in public-user responses.
        created_at:
          type: string
          format: date-time
        profile:
          $ref: '#/components/schemas/Profile'
    Category:
      type: object
      additionalProperties: false
      required: [category_id, name, slug, is_active, created_at]
      properties:
        category_id:
          type: integer
          format: int64
        name:
          type: string
        slug:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
    ListingImage:
      type: object
      additionalProperties: false
      required: [image_id, listing_id, image_url, sort_order, is_primary]
      properties:
        image_id:
          type: integer
          format: int64
        listing_id:
          type: integer
          format: int64
        image_url:
          type: string
          format: uri-reference
          maxLength: 1024
          description: Usually a same-origin `/media/...` URL; legacy administrator imports may be absolute.
        sort_order:
          type: integer
          minimum: 0
        is_primary:
          type: boolean
    Listing:
      type: object
      additionalProperties: false
      required:
        - listing_id
        - seller_id
        - category_id
        - title
        - description
        - brand
        - location
        - price
        - currency
        - item_condition
        - status
        - moderation_status
        - created_at
        - updated_at
        - revision
        - images
      properties:
        listing_id:
          type: integer
          format: int64
        seller_id:
          type: integer
          format: int64
        category_id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        brand:
          type: string
        location:
          type: string
        price:
          type: number
          minimum: 0
        currency:
          type: string
          minLength: 3
          maxLength: 3
        item_condition:
          $ref: '#/components/schemas/ItemCondition'
        status:
          type: string
          enum: [draft, under_review, active, reserved, sold, archived, deleted]
        moderation_status:
          type: string
          enum: [pending, approved, rejected, review, unavailable]
        scam_score:
          type: number
          minimum: 0
          maximum: 1
        scam_label:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        revision:
          type: integer
          format: int64
          minimum: 1
          readOnly: true
          description: Monotonically increasing listing revision used to reject stale concurrent ML updates.
        category:
          $ref: '#/components/schemas/Category'
        seller:
          $ref: '#/components/schemas/User'
        images:
          type: array
          items:
            $ref: '#/components/schemas/ListingImage'
        similarity_score:
          type: number
...
    CartItem:
      type: object
      additionalProperties: false
      required: [listing, quantity]
      properties:
        listing:
          $ref: '#/components/schemas/Listing'
        quantity:
          type: integer
          minimum: 1
          maximum: 99
    Cart:
      type: object
      additionalProperties: false
      required: [cart_id, user_id, items, created_at, updated_at]
      properties:
        cart_id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartItem'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Conversation:
      type: object
      additionalProperties: false
      required: [conversation_id, listing_id, buyer_id, seller_id, created_at, updated_at, members]
      properties:
        conversation_id:
          type: integer
          format: int64
        listing_id:
          type: integer
          format: int64
        buyer_id:
          type: integer
          format: int64
        seller_id:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        members:
          type: array
          items:
            $ref: '#/components/schemas/ConversationMember'
        last_message:
          $ref: '#/components/schemas/Message'
    ConversationMember:
      type: object
      additionalProperties: false
      required: [user, role, joined_at]
      properties:
        user:
          $ref: '#/components/schemas/User'
        role:
          type: string
          enum: [buyer, seller]
        joined_at:
          type: string
          format: date-time
    Message:
      type: object
      additionalProperties: false
      required: [message_id, conversation_id, sender_id, message_text, created_at]
      properties:
        message_id:
          type: integer
          format: int64
        conversation_id:
          type: integer
          format: int64
        sender_id:
          type: integer
          format: int64
        message_text:
          type: string
        created_at:
          type: string
          format: date-time
...
    ListingList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Listing'
...
    ConversationList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Conversation'
    MessageList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Message'
...